==> calaccess-exploration/propositions/00_analysis.ipynb <==
# Propositions analysis

By Ben Welsh

%%capture
%run ./01_props.ipynb
%run ./02_committees.ipynb
%run ./03_filings.ipynb
%run ./04_transfers.ipynb
%run ./05_refunds.ipynb
%run ./06_contributions.ipynb
%run ./07_excel.ipynb
