==> calaccess-exploration/propositions/01_props.ipynb <==
# Propositions

Extract ballot measures from the database and their one-to-many relationship with elections

import os
import settings
import pandas as pd

### Extract all elections

elections_sql = """
SELECT *
FROM "calaccess_processed_election"
INNER JOIN "calaccess_processed_event"
ON ("calaccess_processed_election"."event_ptr_id" = "calaccess_processed_event"."id")
ORDER BY "calaccess_processed_event"."start_time";
"""

elections = pd.read_sql_query(elections_sql, settings.db_connection, parse_dates=["start_time",])

elections.head()

trimmed_elections = elections[['id', 'name', 'start_time']]

trimmed_elections.to_csv(os.path.join(settings.data_dir, "elections.csv"), index=False)

### Extract all propositions

prop_sql = """
SELECT *
FROM calaccess_processed_ballotmeasurecontest as b
INNER JOIN calaccess_processed_contestbase as c
ON b.contestbase_ptr_id = c.id
"""

props = pd.read_sql_query(prop_sql, settings.db_connection)

### Join in the other IDs from the database

prop_calaccess_id = pd.read_sql_table("calaccess_processed_contestidentifier", settings.db_connection)

prop_local_ids = pd.read_sql_table("calaccess_processed_scrapedproposition", settings.db_connection)

prop_ids = prop_calaccess_id.merge(prop_local_ids, left_on="identifier", right_on="scraped_id")[['id_y', 'scraped_id', 'contest_id']].rename(
    columns={
        "id_y": "ccdc_prop_id",
        "scraped_id": "calaccess_prop_id",
        "contest_id": "ocd_prop_id"
    },
)

props = props.merge(prop_ids, left_on="contestbase_ptr_id", right_on="ocd_prop_id")

### Clean up fields and join to elections

trimmed_props = props[[
    'election_id',
    'ocd_prop_id',
    'calaccess_prop_id',
    'ccdc_prop_id',
    'name',
]]

merged_props = pd.merge(
    trimmed_elections,
    trimmed_props,
    how="inner",
    left_on="id",
    right_on="election_id",
    suffixes=["_elections", "_props"]
)

merged_props.rename(columns={
    "name_elections": "election_name",
    "start_time": "election_date",
    "name_props": "prop_name"
}, inplace=True)

merged_props.drop("id", axis=1, inplace=True)

merged_props.to_csv(os.path.join(settings.data_dir, "props.csv"), index=False)

### Filter down to props from the most recent election

recent_props = merged_props[merged_props.election_name == "2016 GENERAL"]

recent_props

recent_props.drop([
    'election_name',
    'election_date',
    'election_id'
], axis=1, inplace=True)

recent_props.to_csv(os.path.join(settings.data_dir, "props-2016.csv"), index=False)
