==> calaccess-exploration/propositions/02_committees.ipynb <==
# Committees

Extract ballot measure committees from the database, as well as their many-to-one links to propositions

import os
import settings
import pandas as pd

### Pull in all ballot measures

props = pd.read_csv(os.path.join(settings.data_dir, "props.csv"))

props.sort_values("ccdc_prop_id").head()

### Pull in all ballot-measure committees

committees = pd.read_sql_table("calaccess_processed_scrapedpropositioncommittee", settings.db_connection)

trimmed_committees = committees[[
    'id',
    'scraped_id',
    'proposition_id',
    'name',
    'position'
]].rename(columns={
    "id": "ccdc_committee_id",
    "scraped_id": "calaccess_committee_id",
    "proposition_id": "ccdc_prop_id",
    "name": "committee_name",
    "position": "committee_position"
})

trimmed_committees.head()

### Merge them together

merged = pd.merge(props, trimmed_committees, how="inner", on="ccdc_prop_id")

merged.head()

merged.to_csv(os.path.join(settings.data_dir, 'committees-and-props.csv'), index=False)

distinct_committees = merged.groupby([
    'calaccess_committee_id',
    'committee_name'
]).size().reset_index()

distinct_committees.rename(columns={0: "prop_count"}, inplace=True)

distinct_committees.to_csv(
    os.path.join(settings.data_dir, 'committees.csv'),
    index=False
)

### Filter down to the most recent election

recent_props = pd.read_csv(os.path.join(settings.data_dir, "props-2016.csv"))

recent_merged = pd.merge(
    recent_props,
    trimmed_committees,
    how="inner",
    on="ccdc_prop_id"
)

recent_merged.head()

recent_merged.to_csv(
    os.path.join(settings.data_dir, 'committees-and-props-2016.csv'),
    index=False
)

recent_distinct_committees = recent_merged.groupby([
    'calaccess_committee_id',
    'committee_name'
]).size().reset_index()

recent_distinct_committees.rename(columns={0: "prop_count"}, inplace=True)

recent_distinct_committees.head()

recent_distinct_committees.to_csv(
    os.path.join(settings.data_dir, 'committees-2016.csv'),
    index=False
)
