==> calaccess-exploration/propositions/03_filings.ipynb <==
# Filings

Extract filings made by ballot measure committees

import os
import settings
import datetime
import pandas as pd

### Pul in all distinct committees

committees = settings.open_csv("committees.csv")

committees.head()

### Pull in all quarterly filings 

form_460_filings = pd.read_sql_table("calaccess_processed_form460filing", settings.db_connection)

form_460_filings.head()

prop_filings = pd.merge(
    form_460_filings,
    committees,
    how="inner",
    left_on="filer_id",
    right_on="calaccess_committee_id"
)

trimmed_filings = prop_filings[[
    'calaccess_committee_id',
    'committee_name',
    'filing_id',
    'date_filed',
    'from_date', 
    'thru_date',
    'total_contributions',
    'total_expenditures_made',
]].sort_values([
    'committee_name',
    'thru_date'
]).rename(columns={
    "filing_id": "calaccess_filing_id",
})

trimmed_filings.total_expenditures_made.sum()

trimmed_filings.to_csv(
    os.path.join(settings.data_dir, 'filings.csv'),
    index=False
)

## 2016 committees

recent_committees = pd.read_csv(
    os.path.join(settings.data_dir, "committees-2016.csv")
)

recent_prop_filings = pd.merge(
    form_460_filings,
    recent_committees,
    how="inner",
    left_on="filer_id",
    right_on="calaccess_committee_id"
)

recent_prop_filings.info()

recent_trimmed_filings = recent_prop_filings[[
    'calaccess_committee_id',
    'committee_name',
    'filing_id',
    'date_filed',
    'from_date', 
    'thru_date',
    'payments_made',
    'loans_made',
    'unpaid_bills',
    'nonmonetary_adjustment',
    'subtotal_cash_payments',
    'total_expenditures_made',
]].sort_values([
    'committee_name',
    'thru_date'
]).rename(columns={
    "filing_id": "calaccess_filing_id",
})

recent_trimmed_filings = recent_trimmed_filings[recent_trimmed_filings.from_date >= datetime.date(2015, 1, 1)]

recent_trimmed_filings.total_expenditures_made.sum()

recent_trimmed_filings.to_csv(
    os.path.join(settings.data_dir, 'filings-2016.csv'),
    index=False
)
