==> calaccess-exploration/propositions/04_transfers.ipynb <==
import os
import settings
import pandas as pd

### All of them

filings = pd.read_csv(os.path.join(settings.data_dir, "filings.csv"))

sched_d = pd.read_sql_table("calaccess_processed_form460scheduleditem", settings.db_connection)

d_items = pd.merge(filings, sched_d, how="inner", left_on="calaccess_filing_id", right_on="filing_id")

d_items.to_csv(os.path.join(settings.data_dir, 'transfers.csv'), index=False)

### 2016

recent_filings = pd.read_csv(os.path.join(settings.data_dir, "filings-2016.csv"))

recent_d_items = pd.merge(recent_filings, sched_d, how="inner", left_on="calaccess_filing_id", right_on="filing_id")

recent_d_items.to_csv(os.path.join(settings.data_dir, 'transfers-2016.csv'), index=False)
