==> calaccess-exploration/propositions/05_refunds.ipynb <==
import os
import settings
import pandas as pd

### All filings

filings = pd.read_csv(os.path.join(settings.data_dir, "filings.csv"))

sql_template = """
SELECT *
FROM "calaccess_processed_form460scheduleeitem"
WHERE filing_id IN ({})
"""

sql = sql_template.format(", ".join(map(str, filings.calaccess_filing_id)))

sched_e = pd.read_sql_query(sql, settings.db_connection)

refunds = sched_e[sched_e.payment_code == 'RFD']

refunds.to_csv(os.path.join(settings.data_dir, "refunds.csv"), index=False)

### 2016

recent_filings = pd.read_csv(os.path.join(settings.data_dir, "filings-2016.csv"))

recent_sql = sql_template.format(", ".join(map(str, recent_filings.calaccess_filing_id)))

recent_sched_e = pd.read_sql_query(recent_sql, settings.db_connection)

recent_refunds = recent_sched_e[recent_sched_e.payment_code == 'RFD']

recent_refunds.to_csv(os.path.join(settings.data_dir, "refunds-2016.csv"), index=False)
