==> calaccess-exploration/propositions/06_contributions.ipynb <==
# Contributions

Extract campaign contributions linked to a committee

import os
import settings
import pandas as pd

### Pull in all filings

filings = pd.read_csv(os.path.join(settings.data_dir, "filings.csv"))

filings.head()

### Pull in all contributions

monetary_sql_template = """
SELECT *
FROM calaccess_processed_form460scheduleaitem
WHERE filing_id IN ({})
"""

monetary_sql = monetary_sql_template.format(
    ",".join(map(str, list(filings.calaccess_filing_id)))
)

contribs = pd.read_sql_query(monetary_sql, settings.db_connection)

trimmed_contribs = contribs.rename(columns={
    'filing_id': 'calaccess_filing_id',
    'id': 'ccdc_contrib_id',
    'transaction_id': 'calaccess_contrib_id'
}).drop([
    'line_item',
    'date_received_thru',
    'transaction_type',
    'memo_reference_number',
    'cumulative_ytd_amount',
    'cumulative_election_amount',
], axis=1)

trimmed_contribs.head()

### Merge them 

merged = pd.merge(
    filings,
    trimmed_contribs,
    how="inner",
    on="calaccess_filing_id"
)

unwanted_columns = [
    'from_date',
    'thru_date',
    'total_contributions',
    'total_expenditures_made',
    'ccdc_contrib_id',
    'date_filed',
    'calaccess_contrib_id',
    'contributor_code',
    'contributor_committee_id',
    'contributor_title',
    'contributor_name_suffix',
    'intermediary_committee_id',
    'intermediary_title',
    'intermediary_lastname',
    'intermediary_firstname',
    'intermediary_name_suffix',
    'intermediary_city',
    'intermediary_state',
    'intermediary_zip',
    'intermediary_employer',
    'intermediary_occupation',
    'intermediary_is_self_employed'
]

trimmed_merged = merged.drop(unwanted_columns, axis=1)

trimmed_merged.head()

trimmed_merged.to_csv(
    os.path.join(settings.data_dir, 'contributions.csv'),
    index=False
)

### Filter down to 2016 contributions

recent_filings = pd.read_csv(os.path.join(settings.data_dir, "filings-2016.csv"))

recent_contribs = trimmed_merged[trimmed_merged.calaccess_filing_id.isin(recent_filings.calaccess_filing_id)]

recent_contribs.head(10)

recent_contribs.to_csv(
    os.path.join(settings.data_dir, 'contributions-2016.csv'),
    index=False
)
