==> calaccess-exploration/propositions/07_expenditures.ipynb <==
# Expenditures

Total up the expenditures by each committee

import os
import settings
import pandas as pd

committees = pd.read_csv(os.path.join(settings.data_dir, 'committees-2016.csv'))

positions = pd.read_csv(os.path.join(settings.data_dir, 'committees-and-props-2016.csv'))[
    ['calaccess_prop_id', 'prop_name', 'calaccess_committee_id', 'committee_position']
]

filings = pd.read_csv(os.path.join(settings.data_dir, "filings-2016.csv"))

merged = committees.merge(
    filings,
    how="inner",
    on="calaccess_committee_id"
)

merged.info()

ctotals = merged.groupby([
    'calaccess_committee_id',
    'committee_name_x'
]).total_expenditures_made.sum().reset_index()

ctotals.rename(columns={"committee_name_x": "committee_name"}, inplace=True)

c2p = ctotals.merge(positions, on="calaccess_committee_id")

c2ptotals = c2p.groupby([
    'prop_name',
    'calaccess_committee_id',
    'committee_name',
    'committee_position'
]).total_expenditures_made.sum().reset_index()

c2ptotals = c2ptotals.sort_values("total_expenditures_made", ascending=False).sort_values("committee_name")

c2ptotals.to_csv(os.path.join(settings.data_dir, 'committee-expenditures-2016.csv'), index=False)

ptotals = c2ptotals.groupby(['prop_name', 'committee_position']).total_expenditures_made.sum().reset_index()

ptotals.to_csv(os.path.join(settings.data_dir, 'proposition-expenditures-2016.csv'), index=False)
