==> calaccess-exploration/propositions/08_excel.ipynb <==
# Excel

Exports the data from this analysis to a Microsoft Excel workbook

import os
import settings
import pandas as pd

### Set up blank Excel workbook

writer = pd.ExcelWriter(
    os.path.join(settings.data_dir, '2016-analysis.xlsx'),
    engine='xlsxwriter'
)

### Propositions

props = settings.open_csv('props-2016.csv')

props.to_excel(writer, sheet_name="Propositions", index=False)

### Committees

committees = settings.open_csv('committees-2016.csv')

committees.to_excel(writer, sheet_name="Committees", index=False)

### Link between committees and propositions

committee_positions = settings.open_csv("committees-and-props-2016.csv")

committee_positions.to_excel(writer, sheet_name="Committee positions", index=False)

### Committee filings and expenditures

filings = settings.open_csv("filings-2016.csv")

filings.to_excel(writer, sheet_name="Filings (Since 2015-01-01)", index=False)

### Transfers

transfers = settings.open_csv("transfers-2016.csv")

transfers.to_excel(writer, sheet_name="Tranfers (Schedule D)", index=False)

### Refunds

refunds = settings.open_csv("refunds-2016.csv")

refunds.to_excel(writer, sheet_name="Refunds (Schedule E)", index=False)

### Save out

writer.save()
